MatchService and PlayerService tests are added

+++ CricketScoreSheetPro.Core/Services/Implementations/MatchService.cs
@@ -1,7 +1,7 @@
-﻿using CricketScoreSheetPro.Core.Helper;
-using CricketScoreSheetPro.Core.Models;
+﻿using CricketScoreSheetPro.Core.Models;
 using CricketScoreSheetPro.Core.Repositories.Interfaces;
 using CricketScoreSheetPro.Core.Services.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -13,22 +13,26 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public MatchService(IRepository<UserMatch> usermatchRepository)
         {
-            _usermatchRepository = usermatchRepository;
+            _usermatchRepository = usermatchRepository ?? throw new ArgumentNullException($"UserMatchRepository is null");
         }
 
         public async Task<UserMatch> AddMatchAsync(UserMatch newmatch)
         {
+            if (newmatch == null) throw new ArgumentNullException($"UserMatch is null");
             var matchAdd = await _usermatchRepository.CreateAsync(newmatch);
-            return matchAdd.Object;
+            return matchAdd;
         }
 
         public async Task UpdateMatchAsync(string matchId, UserMatch updateMatch)
         {
+            if (updateMatch == null) throw new ArgumentNullException($"UserMatch is null");
+            if (string.IsNullOrEmpty(matchId)) throw new ArgumentException($"Match ID is null");
             await _usermatchRepository.CreateWithIdAsync(matchId, updateMatch);
         }
 
         public async Task<UserMatch> GetMatchAsync(string matchId)
         {
+            if (string.IsNullOrEmpty(matchId)) throw new ArgumentException($"Match ID is null");
             var team = await _usermatchRepository.GetItemAsync(matchId);
             return team;
         }
@@ -36,7 +40,7 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
         public async Task<IList<UserMatch>> GetMatchesAsync()
         {
             var matches = await _usermatchRepository.GetListAsync();
-            return Common.ConvertFirebaseObjectCollectionToList(matches);
+            return matches;
         }
 
         public async Task DeleteAllMatchesAsync()
@@ -46,6 +50,7 @@ namespace CricketScoreSheetPro.Core.Services.Implementations
 
         public async Task DeleteMatchAsync(string matchId)
         {
+            if (string.IsNullOrEmpty(matchId)) throw new ArgumentException($"Match ID is null");
             await _usermatchRepository.DeleteByIdAsync(matchId);
         }
     }
